Fix authentication issues

+++ deploy.sh
@@ -2,7 +2,7 @@
 
 function syntax()
 {
-    echo "Syntax: $(basename $0) src [localhost|buckets ...]"
+    echo "Syntax: $(basename $0) [-f|--force] src [localhost|buckets ...]"
 }
 
 if [ $# -lt 1 ]; then
@@ -11,7 +11,12 @@ if [ $# -lt 1 ]; then
 elif [ "$1" == "-h" -o "$1" == "--help" -o "$1" == "help" ]; then
     syntax
     exit 0
-elif [ ! -d "$1" ]; then
+elif [ "$1" == "-f" -o "$1" == "--force" ]; then
+    FORCE=yes
+    shift 1
+fi
+
+if [ ! -d "$1" ]; then
     echo "ERROR: $(basename $0) folder '$1' not found" > /dev/stderr
     exit 2
 fi
@@ -28,14 +33,14 @@ else
     shift 1
     while [ $# -ge 1 ]; do
         for file in $(ls) ; do
-            if [ -f "$file" ]; then
+            if [ -f "$file" -a -z "$(grep ^$file .deploy-ignore 2>/dev/null)" ]; then
                 target="s3://$1/$file"
-                if [ -z "$(grep ^$file .deploy-once)" -o -z "$(aws s3 ls $1/$file)" ]; then
+                if [ "${FORCE:-no}" == "yes" -o -z "$(grep ^$file .deploy-once)" -o -z "$(aws s3 ls $1/$file)" ]; then
                     # echo $file
                     aws s3 cp "$file" "$target"
                     aws s3api put-object-acl --bucket "$1" --key "$file" --acl public-read
                 else
-                    echo "once  : ./$file to $target"
+                    echo "skipped: ./$file to $target"
                 fi
             fi
         done
